requirements.txt
# Core Dependencies
pydantic>=2.0.0
pyyaml>=6.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
cryptography>=41.0.0
tenacity>=8.2.0
httpx>=0.24.0

# UI
PyQt6>=6.5.0

# ML & Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Agent APIs
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
slack-sdk>=3.21.0

# Voice
pvporcupine>=2.2.0
openai-whisper>=20230314
piper-tts>=1.2.0
pyaudio>=0.2.13
sounddevice>=0.4.6

# LLM
ollama>=0.1.0
